+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/artwork/service/ArtworkService.java
@@ -6,7 +6,6 @@ import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkSaveRequest;
 import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkUpdateRequest;
 import com.startingblue.fourtooncookie.artwork.exception.ArtworkDuplicateException;
 import com.startingblue.fourtooncookie.artwork.exception.ArtworkNotFoundException;
-import com.startingblue.fourtooncookie.global.config.XmlMessageSource;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
@@ -23,7 +22,7 @@ import java.util.Objects;
 public class ArtworkService {
 
     private final ArtworkRepository artworkRepository;
-    private final XmlMessageSource messageSource;
+    private final MessageSource xmlMessageSource;
 
     public void createArtwork(ArtworkSaveRequest request) {
         verifyUniqueArtwork(request.title(), request.thumbnailUrl());
@@ -70,7 +69,7 @@ public class ArtworkService {
     }
 
     public String getLocalizedArtworkTitle(Long artworkId, Locale locale) {
-        return Objects.requireNonNull(messageSource.resolveCode("artwork.name." + artworkId, locale)).format(null);
+        return Objects.requireNonNull(xmlMessageSource.getMessage("artwork.name." + artworkId, null, locale));
     }
 
 }
