+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/service/CharacterService.java
@@ -5,13 +5,13 @@ import com.startingblue.fourtooncookie.artwork.service.ArtworkService;
 import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.character.domain.CharacterRepository;
 import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
-import com.startingblue.fourtooncookie.global.config.XmlMessageSource;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import com.startingblue.fourtooncookie.character.dto.request.CharacterSaveRequest;
 import com.startingblue.fourtooncookie.character.dto.request.CharacterUpdateRequest;
 import com.startingblue.fourtooncookie.character.exception.CharacterDuplicateException;
 import com.startingblue.fourtooncookie.character.exception.CharacterNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,8 +26,7 @@ public class CharacterService {
 
     private final CharacterRepository characterRepository;
     private final ArtworkService artworkService;
-
-    private final XmlMessageSource messageSource;
+    private final MessageSource xmlMessageSource;
 
     public void createCharacter(final CharacterSaveRequest request) {
         CharacterVisionType visionType = findByCharacterVisionType(request.characterVisionType());
@@ -117,7 +116,7 @@ public class CharacterService {
     }
 
     public String getLocalizedCharacterName(Long characterId, Locale locale) {
-        return Objects.requireNonNull(messageSource.resolveCode("character.name." + characterId, locale)).format(null);
+        return Objects.requireNonNull(xmlMessageSource.getMessage("character.name." + characterId, null, locale));
     }
 
 }
