+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/global/config/MessageConfig.java
@@ -1,5 +1,6 @@
 package com.startingblue.fourtooncookie.global.config;
 
+import com.startingblue.fourtooncookie.global.domain.XmlMessageSource;
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -8,7 +9,7 @@ import org.springframework.context.annotation.Configuration;
 public class MessageConfig {
 
     @Bean
-    public MessageSource messageSource() {
+    public MessageSource xmlMessageSource() {
         return new XmlMessageSource();
     }
 }
